tests/requirements.txt
coverage==4.3.1
docutils==0.13.1
flake8==3.2.1
ghp-import==0.5.4
pycodestyle==2.2.0
pyflakes==1.4.0
pytest==3.0.5
pytest-aiohttp==0.1.3
pytest-cov==2.4.0
pytest-isort==0.1.0
pytest-mock==1.5.0
pytest-sugar==0.8.0
pytest-toolbox==0.1
pytest-xdist==1.15.0
Sphinx==1.5.1

# required to run "start" apps
aiohttp-jinja2==0.13.0
aiohttp-session[secure]==0.8.0
aiohttp-session[aioredis]==0.8.0
aiopg==0.13.0
SQLAlchemy==1.1.4
trafaret-config==1.0.1
